hibernate frame set up!

+++ MSHServerModule/src/main/java/datahelper/HibernateHelper.java
@@ -1,27 +1,33 @@
 package datahelper;
 
-import org.hibernate.Criteria;
-import org.hibernate.HibernateException;
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
+import org.hibernate.*;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.criterion.Restrictions;
+import org.hibernate.exception.ConstraintViolationException;
 import util.ResultMessage;
-
+import java.util.ArrayList;
 import java.util.List;
 
 /**
  * Created by SilverNarcissus on 16/11/12.
+ * DataHelper which use hibernate frame
  */
-public class HibernateHelper implements DataHelper{
+public class HibernateHelper implements DataHelper {
 
     private SessionFactory sessionFactory;
     private Session session;
     private String classFullName;
 
-    private HibernateHelper(String cfgLocation,String classFullName) {
+    public HibernateHelper(String cfgLocation) {
         Configuration configuration = new Configuration();
         sessionFactory = configuration.configure(cfgLocation).buildSessionFactory();
+    }
+
+    /**
+     * 设置PO类的全名
+     * @param classFullName
+     */
+    public void setClassName(String classFullName){
         this.classFullName=classFullName;
     }
 
@@ -29,7 +35,7 @@ public class HibernateHelper implements DataHelper{
      * 初始化Session
      */
     private void setUpSession() {
-        Session session = sessionFactory.openSession();
+        session = sessionFactory.openSession();
         session.beginTransaction();
     }
 
@@ -45,18 +51,20 @@ public class HibernateHelper implements DataHelper{
      * 将对象添加到数据库表中
      *
      * @param o
+     * @since 1.6
+     * @return 保存操作的结果信息
      */
-    public ResultMessage save(Object o) throws HibernateException{
+    public ResultMessage save(Object o) {
         try {
             setUpSession();
             session.contains(o);
             session.save(o);
             commitAndClose();
-        }
-        catch (S){
-
-        }
-        catch (Exception e){
+        } catch (ConstraintViolationException e) {
+            e.printStackTrace();
+            return ResultMessage.EXIST;
+        } catch (Exception e) {
+            e.printStackTrace();
             return ResultMessage.FAILED;
         }
         return ResultMessage.SUCCESS;
@@ -64,25 +72,43 @@ public class HibernateHelper implements DataHelper{
 
     /**
      * 更新数据表中的对象
-     *
      * @param o
+     * @return 更新操作的结果信息
      */
-    public ResultMessage update(Object o) throws HibernateException{
-        setUpSession();
-        session.update(o);
-        commitAndClose();
+    public ResultMessage update(Object o) {
+        try {
+            setUpSession();
+            session.update(o);
+            commitAndClose();
+        } catch (StaleStateException e) {
+            e.printStackTrace();
+            return ResultMessage.NOT_EXIST;
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResultMessage.FAILED;
+        }
+        return ResultMessage.SUCCESS;
     }
+
 
     /**
      * 删除数据表中的对象
+     *
      * @param ID
-     * @throws ClassNotFoundException
-     * @throws HibernateException
      */
-    public ResultMessage delete(String ID) throws ClassNotFoundException, HibernateException {
-        setUpSession();
-        session.delete(exactlyQuery("ID", ID));
-        commitAndClose();
+    public ResultMessage delete(String ID) {
+        try {
+            setUpSession();
+            if (exactlyQuery("ID", ID) == null) {
+                return ResultMessage.NOT_EXIST;
+            }
+            session.delete(exactlyQuery("ID", ID));
+            commitAndClose();
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResultMessage.FAILED;
+        }
+        return ResultMessage.SUCCESS;
     }
 
     /**
@@ -94,10 +120,15 @@ public class HibernateHelper implements DataHelper{
      * @throws ClassNotFoundException
      * @throws HibernateException
      */
-    public Object exactlyQuery(String field, Object value) throws ClassNotFoundException, HibernateException {
-        Criteria criteria = SetUpCriteria();
-        criteria.add(Restrictions.eq(field, value));
-        return (criteria.list().size() != 0) ? null : criteria.list().get(0);
+    public Object exactlyQuery(String field, Object value) {
+        try {
+            Criteria criteria = SetUpCriteria();
+            criteria.add(Restrictions.eq(field, value));
+            return (criteria.list().size() != 0) ? null : criteria.list().get(0);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
     }
 
     /**
@@ -106,13 +137,11 @@ public class HibernateHelper implements DataHelper{
      * @param field 属性名称
      * @param value 属性值
      * @return PO列表
-     * @throws ClassNotFoundException
-     * @throws HibernateException
+     *
      */
-    public List<Object> prefixMatchSearch(String field, String value) throws ClassNotFoundException, HibernateException {
-        Criteria criteria = SetUpCriteria();
-        criteria.add(Restrictions.like(field, value + "%"));
-        return criteria.list();
+    public List<Object> prefixMatchQuery(String field, String value) {
+        value = value + "%";
+        return likePatternQuery(field, value);
     }
 
     /**
@@ -121,13 +150,11 @@ public class HibernateHelper implements DataHelper{
      * @param field 属性名称
      * @param value 属性值
      * @return PO列表
-     * @throws ClassNotFoundException
-     * @throws HibernateException
+     *
      */
-    public List<Object> suffixMatchQuery(String field, String value) throws ClassNotFoundException, HibernateException {
-        Criteria criteria = SetUpCriteria();
-        criteria.add(Restrictions.like(field, "%" + value));
-        return criteria.list();
+    public List<Object> suffixMatchQuery(String field, String value) {
+        value = "%" + value;
+        return likePatternQuery(field, value);
     }
 
     /**
@@ -136,13 +163,11 @@ public class HibernateHelper implements DataHelper{
      * @param field
      * @param value
      * @return PO列表
-     * @throws ClassNotFoundException
-     * @throws HibernateException
+     *
      */
-    public List<Object> FuzzyQuery(String field, String value) throws ClassNotFoundException, HibernateException {
-        Criteria criteria = SetUpCriteria();
-        criteria.add(Restrictions.like(field, "%" + value + "%"));
-        return criteria.list();
+    public List<Object> FuzzyQuery(String field, String value) {
+        value = "%" + value + "%";
+        return likePatternQuery(field, value);
     }
 
     /**
@@ -152,13 +177,17 @@ public class HibernateHelper implements DataHelper{
      * @param min
      * @param max
      * @return PO列表
-     * @throws ClassNotFoundException
-     * @throws HibernateException
+     *
      */
-    public List<Object> RangeQuery(String field, Object min, Object max) throws ClassNotFoundException, HibernateException {
-        Criteria criteria = SetUpCriteria();
-        criteria.add(Restrictions.between(field, min, max));
-        return criteria.list();
+    public List<Object> RangeQuery(String field, Object min, Object max) {
+        try {
+            Criteria criteria = SetUpCriteria();
+            criteria.add(Restrictions.between(field, min, max));
+            return criteria.list();
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ArrayList<Object>();
+        }
     }
 
     /**
@@ -168,6 +197,24 @@ public class HibernateHelper implements DataHelper{
      * @throws ClassNotFoundException
      */
     private Criteria SetUpCriteria() throws ClassNotFoundException {
-        return session.createCriteria(Class.forName("hotelPO"));
+        return session.createCriteria(Class.forName(classFullName));
+    }
+
+    /**
+     * 利用模糊查找返回符合条件的PO列表
+     *
+     * @param field
+     * @param value
+     * @return PO列表
+     */
+    private List<Object> likePatternQuery(String field, String value) {
+        try {
+            Criteria criteria = SetUpCriteria();
+            criteria.add(Restrictions.like(field, value));
+            return criteria.list();
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ArrayList<Object>();
+        }
     }
 }
